% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlCap.R
\name{xlCap}
\alias{xlCap}
\title{Capitalising words}
\usage{
xlCap(s, strict = FALSE)
}
\arguments{
\item{s}{the string to convert.}

\item{strict}{logical, if \code{True} then all words will have only
the first letter capitalised and others in lower cases.}
}
\description{
A sophisticated version of word capitalisation
}
