% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlConstants.R
\name{xlCRS}
\alias{xlCRS}
\title{Coordinate Reference System objects}
\usage{
xlCRS(system = "wgs84")
}
\arguments{
\item{system}{A string, either \code{'wgs84'} (default), \code{'bng'} or \code{'google'}.}
}
\value{
A CRS (coordinate reference system) object from calling \code{\link{CRS}}.
  So far the elements are the CRS objects
\code{wgs84}, \code{bng} and \code{google}. More will come as needed.
}
\description{
Returns a CRS object for geo-spatial analysis.
}
\details{
There are a few constants that are useful in geo-spatial analysis: CRS
such as \href{https://en.wikipedia.org/wiki/World_Geodetic_System}{World
Geodetic System (wgs84)} and \href{https://en.wikipedia.org/wiki/Ordnance_Survey_National_Grid}{
British National Grid (bng)}.

Input \code{system} can be one of the below values:
\itemize{
  \item \code{wgs84}: \href{http://spatialreference.org/ref/epsg/4326/}{
  \code{EPSG:4326}}, WGS 84 aka the latest version of
  World Geodetic System (1984), with the generic latitude and
  longitude measured degrees.
  \item \code{bng}: \href{http://spatialreference.org/ref/epsg/27700/}{
  \code{EPSG:27700}}, British National Grid, with easting and
  northing measured in meters.
  \item \code{google}: \href{http://wiki.openstreetmap.org/wiki/EPSG:3857}{
  \code{EPSG:3857}}, the Web Mercator project, aka the Google
  projection. Sometimes the reference is also \code{EPSG:900913}. If it's
  not obvious, squint and look again. Still no? \code{900913} spells \code{Google}.
  }
}
\references{
\url{http://spatialreference.org/ref/epsg/}.
}
