% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlConstants.R
\name{xlEarthRadius}
\alias{xlEarthRadius}
\title{Actual radius of the earth}
\source{
The numbers and equations are taken from
\href{http://en.wikipedia.org/wiki/Figure_of_the_Earth}{Wikipedia}
}
\usage{
xlEarthRadius(latitude)
}
\arguments{
\item{latitude}{numerical, the latitude in degrees.}
}
\value{
the Earth's radius in meters
}
\description{
This function calculates the radius of the earth at any latitude
}
