% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlExportCSV.R
\name{xlExportCSV}
\alias{xlExportCSV}
\title{Export multiple dataframes to local CSVs}
\usage{
xlExportCSV(dfs, file_loc, file_names, suffix = xlTxtDate(), ...)
}
\arguments{
\item{dfs}{A list of dataframes to be exported.}

\item{file_loc}{Single character string for the path to which the files are to be exported.}

\item{file_names}{Character vector, the file names to be saved.}

\item{suffix}{A character vector to append to the end of the file names (such as the current date),}

\item{...}{passed to \code{\link{write.csv}}.}
}
\description{
This function uses write.csv to export one or more dataframes to CSV file(s)
on the local drive.
}
\examples{
\dontrun{
xlExportCSV(list(df1, df2), './output', c('table1', 'table2'))
}


}
