% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlGGRadar.R
\name{xlGGRadar}
\alias{xlGGRadar}
\title{xlGGRadar}
\usage{
xlGGRadar(plot.data, font.radar = "Circular Air Light",
  values.radar = c("0\%", "50\%", "100\%"),
  axis.labels = colnames(plot.data)[-1], grid.min = 0, grid.mid = 0.5,
  grid.max = 1, centre.y = grid.min - ((1/9) * (grid.max - grid.min)),
  plot.extent.x.sf = 1, plot.extent.y.sf = 1.2, x.centre.range = 0.02 *
  (grid.max - centre.y), label.centre.y = FALSE, grid.line.width = 0.5,
  gridline.min.linetype = "longdash", gridline.mid.linetype = "longdash",
  gridline.max.linetype = "longdash", gridline.min.colour = "grey",
  gridline.mid.colour = "#007A87", gridline.max.colour = "grey",
  grid.label.size = 7, gridline.label.offset = -0.1 * (grid.max - centre.y),
  label.gridline.min = TRUE, axis.label.offset = 1.15,
  axis.label.size = 8, axis.line.colour = "grey", group.line.width = 1.5,
  group.point.size = 6, background.circle.colour = "#D7D6D1",
  background.circle.transparency = 0.2, plot.legend = if (nrow(plot.data) >
  1) TRUE else FALSE, legend.title = "", plot.title = "",
  legend.text.size = grid.label.size)
}
\arguments{
\item{plot.data}{dataframe comprising one row per group (cluster); col1 = \emph{group name}, cols2 - n = \emph{variable values}}

\item{axis.labels}{names of axis labels if other than column names supplied via \code{plot.data}, default \code{colnames(plot.data)[-1]}}

\item{grid.min}{value at which mininum grid line is plotted, default \code{-0.5}}

\item{grid.mid}{value at which 'average' grid line is plotted, default \code{0}}

\item{grid.max}{value at which 'average' grid line is plotted, default \code{0.5}}

\item{centre.y}{value of y at centre of plot, default < \code{grid.min}}

\item{plot.extent.x.sf}{controls relative size of plot horizontally, default \code{1.2}}

\item{plot.extent.y.sf}{controls relative size of plot vertically, default \code{1.2}}

\item{x.centre.range}{controls axis label alignment. Default behaviour is to left-align axis labels on left hand side of
plot (\code{x < -x.centre.range}); right-align labels on right hand side of plot (\code{x > +x.centre.range});
and centre align those labels for which \code{-x.centre.range < x < +x.centre.range}}

\item{label.centre.y}{whether value of y at centre of plot should be labelled, default \code{FALSE}}

\item{grid.line.width}{Default \code{0.5}}

\item{gridline.min.linetype}{Default \code{'longdash'}}

\item{gridline.mid.linetype}{Default \code{'longdash'}}

\item{gridline.max.linetype}{Default \code{'longdash'}}

\item{gridline.min.colour}{Default \code{'grey'}}

\item{gridline.mid.colour}{Default \code{'blue'}}

\item{gridline.max.colour}{Default \code{'grey'}}

\item{grid.label.size}{text size, default \code{4}}

\item{gridline.label.offset}{displacement to left/right of central vertical axis,
default \code{-0.02 * (grid.max - centre.y)}}

\item{label.gridline.min}{whether or not to label the mininum gridline, default \code{TRUE}}

\item{axis.label.offset}{vertical displacement of axis labels from maximum grid line, measured relative to
circle diameter, default \code{1.15}}

\item{axis.label.size}{text size, default \code{3}}

\item{axis.line.colour}{line colour, default \code{grey}}

\item{group.line.width}{Default \code{1}}

\item{group.point.size}{Default \code{4}}

\item{background.circle.colour}{Default \code{'yellow'}}

\item{background.circle.transparency}{Default \code{0.2}}

\item{plot.legend}{whether to include a plot legend, default = \code{FALSE} for one cluster; \code{TRUE} for 2+ clusters}

\item{legend.title}{Default \code{'Cluster'}}

\item{legend.text.size}{Default \code{grid.label.size = 4}}
}
\description{
This function is taken from the \code{\href{https://github.com/ricardo-bion/ggradar}{ggradar}} package.
}
\details{
ggradar allows you to build radar charts with ggplot2. This package is based on
\href{http://rstudio-pubs-static.s3.amazonaws.com/5795_e6e6411731bb4f1b9cc7eb49499c2082.html}{Paul Williamson}'s code,
with new aesthetics and compatibility with ggplot2 2.0.
It was inspired by \href{http://www.buildingwidgets.com/blog/2015/12/9/week-49-d3radarr}{d3radaR},
an htmlwidget built by \href{https://github.com/timelyportfolio}{timelyportfolio}.

\strong{Plot extent} - parameters to rescale the extent of the plot vertically and horizontally, in order to
allow for ggplot default settings placing parts of axis text labels outside of plot area.
Scaling factor is defined relative to the circle diameter (\code{grid.max - centre.y}).

\strong{Grid lines} - includes separate controls for the appearance of some aspects the 'minimum', 'average' and 'maximum' grid lines.
}
\references{
Most of the code and documentation is from
\url{http://rstudio-pubs-static.s3.amazonaws.com/5795_e6e6411731bb4f1b9cc7eb49499c2082.html},
and the function code is taken from \url{https://github.com/ricardo-bion/ggradar}.
}
\author{
Ricardo Bion
}
