% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlHelpers.R
\name{xlTxtDate}
\alias{xlTxtDate}
\alias{xlPct}
\alias{xlAddCommas}
\alias{xlSQLIn}
\title{Tiny helpers}
\usage{
xlTxtDate(date = Sys.Date(), date_format = "\%Y\%m\%d")

xlPct(num, digits = 2, num_format = "f", ...)

xlAddCommas(num, big.mark = ",", num_format = "d",
  preserve.width = "none", ...)

xlSQLIn(in_values)
}
\arguments{
\item{date}{A date vector to be formated, default \code{Sys.Date()} for today's date.}

\item{date_format}{A format string for date, default \code{'\%Y\%m\%d'}
yielding e.g. \code{20171231}.}

\item{num}{Numeric vector, to be transformed into string vector of various formats.}

\item{digits}{Integer for the number of digits after decimal places in \code{xlPct}
function.}

\item{num_format}{Single length string for numerical conversion format
in \code{xlPct} and \code{xlAddCommas}.}

\item{...}{Additional arguments passed to \code{xlPct}, \code{xlAddCommas} etc..}

\item{big.mark}{the symbol for separating digits, using \code{','} as the default.}

\item{preserve.width}{parameters for \code{xlAddCommas} for saving additional
spaces for output string.}

\item{in_values}{A vector of any kind, to be transformed into a string separated by
commas for an 'in statement' as part of a SQL query.}
}
\description{
One Liner helper functions.
}
\details{
These functions typically take various types of vectors and return
transformed vectors / strings with specific formats that are common
in day to day daa manipulations.
}
\examples{
xlTxtDate()
xlPct(0.83)
xlAddCommas(93976)
xlSQLIn(c('Display', 'Organic'))


}
