% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlPalettes.R
\name{xlColPal}
\alias{xlColPal}
\alias{xlColPalDisp}
\title{Colour Palettes}
\usage{
xlColPal(cols = NULL, palette = "wipro", alpha = 1)

xlColPalDisp(cols = NULL, palette = "wipro", ...)
}
\arguments{
\item{cols}{character or numeric vector for the names or the indices of colours
in the palette specified by \code{palette}.}

\item{palette}{a string, the colour palette name.}

\item{alpha}{a number between \code{[0, 1]} for the opacity, 0 being opac.}

\item{...}{parameters passed to either \code{\link{rgb}} for \code{\link{xlColPal}},
or \code{\link{img}} for \code{\link{xlColPalDisp}}.}
}
\description{
These functions return hexadecimal colours from the
various colour palettes I have collected over time from
places including companies I have worked for.
}
\details{
To do: how to use.

Currently the available palettes are as follow:
\itemize{
 \item \code{"ee"}: EE colours, from the official internal branding document.
 Internally the colours are in the list \code{ee_hex_cols} in hexadecimal format.
 \item \code{"wipro"}: Wipro colours, from the initial template (see Reference below).
 Internally the colours are in the list \code{wipro_rgb_cols} in RGB format.
}
}
\examples{
xlColPal()
xlColPal(c(1, 2, 3))
xlColPal(c('red', 'cyan', 'grey'))
xlColPal(rep(c(1, 2), 3)) # repeats are allowed
xlColPalDisp()
xlColPalDisp(c(1, 2, 3), 'ee')

\dontrun{
# What not to do:
xlColPal(c(1, 2, 3), palette = 'wrong') # error, wrong palette name
xlColPal(c(356, 6))                     # error, as 356 is outside of 1:10
xlColPal(c(1, 'cyan', 'grey'))          # error, mixed cols element types
xlColPal(c('1'))                        # error, numeric should not be quoted
}
}
\references{
Wipro colours are taken from the internal template https://wipro365-my.sharepoint.com/personal/scshekar_wipro_com/Documents/Analytics%20-%20Overview%20-%202017.pptx?web=1
}
