% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlReadScript.R
\name{xlReadScript}
\alias{xlReadScript}
\title{xlReadScript}
\usage{
xlReadScript(script, cmt_sym = "--", ...)
}
\arguments{
\item{script}{A \code{\link{connections}} object or a character string,
exactly the same input expected by \code{\link{readLines}} function.}

\item{cmt_sym}{The character sequence that signifies comment
strings, default \code{'--'} for SQL.}

\item{...}{passed to \code{\link{readLines}}.}
}
\description{
Read standalone script into R as a single character string.
}
\details{
The funtion calls \code{\link{readLines}} to read in the script text file,
strips out comments blocks, inline comments, and merge all lines into
one single string. As it stands, it works best for a fairly simple
SQL script. I will be albe to expand the functionality once
I'm exposed to more complex script. This function is suitable for reading
script that does not use line-by-line execution, but rather entirely
relies on keywords, e.g. SQL.
}
\examples{
\dontrun{
# postgres
tbl(db, sql(xlReadScript('~/Script/script1.sql')))

# MS SQL Server
sqlQuery(conn, xlReadScript('~/Script/script2.sql'))
}

}
