% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlPalettes.R
\name{xlRgbtoHex}
\alias{xlRgbtoHex}
\title{RGB to Hex colour
Convert RGB colours to Hexadecimal format}
\usage{
xlRgbtoHex(rgb_list, max_col_val = 255, col_names = names(rgb_list), ...)
}
\arguments{
\item{rgb_list}{a list of RGB numerical vectors.}

\item{max_col_val}{numerical, the maximum value for the RGB colours, default 255}

\item{col_names}{character vector, the names of the colours, default the names of the input list}
}
\description{
RGB to Hex colour
Convert RGB colours to Hexadecimal format
}
\keyword{internal}
