% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlWC.R
\name{xlWC}
\alias{xlWC}
\title{Week commencing date}
\usage{
xlWC(date = Sys.Date(), pre_wk = 0, wk_start = 1, format = NULL)
}
\arguments{
\item{date}{Date value for which the week start date can be found.
Default is \code{\link{Sys.Date}}}

\item{pre_wk}{Integer for the number of weeks prior to the lookup date.
Default is 0, no shifting of the week.}

\item{wk_start}{Integer, possible values are 1 - 7, correspnoding to Monday to Sunday as
the start of the week. Default is 1 for Monday.}

\item{format}{A string for date output format, couuld be e.g. \code{'\%Y\%m\%d'}.}
}
\description{
Find the week commencing of a date, given a specific day as the start of week.
}
\examples{
# Get the close Monday before today
xlWC()

# Get the Sunday before 2 weeks prior to 1st Jan 2017
xlWC(as.Date('2017-01-01'), pre_wk = 2, wk_start = 7)


}
